Fix pools stats when the results have multiple users per DB.
If you're using auth_query mode + autodb for pgbouncer, you wind up having
multiple users for a single DB.  That output looks roughly like thus:

```
 database  |   user    | cl_active | cl_waiting | sv_active | sv_idle | sv_used | sv_tested | sv_login | maxwait | maxwait_us | pool_mode
-----------+-----------+-----------+------------+-----------+---------+---------+-----------+----------+---------+------------+-----------
 paste     | paste     |         0 |          0 |         0 |       2 |       0 |         0 |        0 |       0 |          0 | session
 paste     | pgbouncer |         0 |          0 |         0 |       2 |       0 |         0 |        0 |       0 |          0 | session
```

With how the code was written, this would break cardinality.  User *must* be
a label in this case.  As for pool_mode, it doesn't have to be a label- it's
just useful for exporting it.

This fixes ticket #1 and #7 (same issue)