Allow pgbouncer exporter to start even if the backend is down.
Previous code required the backend to be up for the exporter to start;
now that pgbouncer_up is properly implemented, we can start the exporter
even if the backend is down- we just report that it's down via metrics.

This removes a gotcha for deploys where the exporter is started before
the backend is up; it's rare, but it's fixable, thus I'm fixing it.

Note: this requires pulling in a new version of lib/pg that has NewConnector
support.

Fixes #9